refactor!: split source files
Splits up and decouples some of the codebase in order to improve
maintainability, testing and extension support.

If you're using a custom parser, validator or resolver with requires
from this plugin then this is a BREAKING CHANGE. Update your require
path to `nvim_context_vt.utils`.
