+++ doc/nvim_context_vt.txt
@@ -49,7 +49,9 @@ To customize the behavior use the setup function:
 
       -- Custom virtual text node parser callback
       -- Default: nil
-      custom_parser = function(node, ft, ts_utils)
+      custom_parser = function(node, ft, opts)
+        local ts_utils = require('nvim-treesitter.ts_utils')
+
         -- If you return `nil`, no virtual text will be displayed.
         if node:type() == 'function' then
           return nil
@@ -60,9 +62,9 @@ To customize the behavior use the setup function:
 
       -- Custom node validator callback
       -- Default: nil
-      custom_validator = function(node, ft)
+      custom_validator = function(node, ft, opts)
         -- Internally a node is matched against min_rows and configured targets
-        local default_validator = require('nvim_context_vt').default_validator
+        local default_validator = require('nvim_context_vt.utils').default_validator
         if default_validator(node, ft) then
           -- Custom behaviour after using the internal validator
           if node:type() == 'function' then
@@ -74,7 +76,7 @@ To customize the behavior use the setup function:
 
       -- Custom node virtual text resolver callback
       -- Default: nil
-      custom_resolver = function(nodes, ft)
+      custom_resolver = function(nodes, ft, opts)
         -- By default the last node is used
         return nodes[#nodes]
       end,
