--- lua/nvim_context_vt.lua
@@ -1,268 +1,0 @@
-local ts_utils = require('nvim-treesitter.ts_utils')
-local parsers = require('nvim-treesitter.parsers')
-
-local ns = vim.api.nvim_create_namespace('context_vt')
-
-local opts = {
-    min_rows = 1,
-    custom_parser = nil,
-    custom_validator = nil,
-    custom_resolver = nil,
-    highlight = 'ContextVt',
-    disable_ft = {},
-    disable_virtual_lines = false,
-    disable_virtual_lines_ft = {},
-    prefix = '-->',
-}
-
-local targets = {
-    'function',
-    'method_declaration',
-    'function_declaration',
-    'function_definition',
-    'lexical_declaration',
-    'local_function',
-    'arrow_function',
-    'method_definition',
-    'arguments',
-
-    'try_statement',
-    'catch_clause',
-    'finally_clause',
-
-    'if_statement',
-    'if_expression',
-
-    'switch_expression',
-
-    'class_definition',
-    'class_declaration',
-    'struct_expression',
-
-    'test_expression',
-
-    'while_expression',
-    'while_statement',
-
-    'for_expression',
-    'foreach_statement',
-    'for_statement',
-    'for_in_statement',
-
-    'with_statement',
-
-    -- rust
-    'match_expression',
-    'if_let_expression',
-    'tuple_struct_pattern',
-    'while_let_expression',
-    'loop_expression',
-    'function_item',
-    'struct_item',
-    'unsafe_block',
-
-    -- ruby target
-    'class',
-    'module',
-    'method',
-    'do_block',
-    'if',
-    'while',
-    'for',
-
-    -- typescript/javascript
-    'interface_declaration',
-    'enum_declaration',
-    'call_expression',
-    'export_statement',
-    'property_signature',
-
-    -- lua,
-    'local_variable_declaration',
-    'variable_declaration',
-
-    -- go
-    'type_declaration',
-    'type_spec',
-    'short_var_declaration',
-    'defer_statement',
-    'expression_switch_statement',
-    'composite_literal',
-    'element',
-    'func_literal',
-    'go_statement',
-    'select_statement',
-    'communication_case',
-    'default_case',
-
-    -- cpp
-    'case_statement',
-    'for_range_loop',
-
-    -- python
-    'list',
-    'pair',
-    'assignment',
-    'decorated_definition',
-
-    -- yaml
-    'block_mapping_pair',
-}
-
-local ignore_root_targets = {
-    'program',
-    'document',
-}
-
-local line_targets = {
-    'function_definition',
-    'class_definition',
-    'if_statement',
-    'try_statement',
-    'with_statement',
-    'for_statement',
-
-    -- python
-    'decorated_definition',
-
-    -- yaml
-    'block_mapping_pair',
-}
-
-local line_ft = {
-    'python',
-    'yaml',
-}
-
-local function default_parser(node)
-    return opts.prefix .. ' ' .. ts_utils.get_node_text(node, 0)[1]
-end
-
-local function default_validator(node)
-    if vim.tbl_contains(targets, node:type()) then
-        return node:end_() > node:start() + opts.min_rows
-    end
-    return false
-end
-
-local function default_resolver(nodes)
-    return nodes[#nodes]
-end
-
-local function find_virtual_text_nodes(validator, ft)
-    local result = {}
-    local node = ts_utils.get_node_at_cursor()
-
-    while node ~= nil and not vim.tbl_contains(ignore_root_targets, node:type()) do
-        if validator(node, ft) then
-            local target_line = node:end_()
-            if not result[target_line] then
-                result[target_line] = {}
-            end
-            table.insert(result[target_line], node)
-        end
-
-        node = node:parent()
-    end
-
-    return result
-end
-
-local function create_virtual_text_factory(parser, ft)
-    local is_line_ft = vim.tbl_contains(line_ft, ft)
-    local skip_line_ft = opts.disable_virtual_lines or vim.tbl_contains(opts.disable_virtual_lines_ft, ft)
-
-    local function lines_from_nodes(nodes)
-        local lines = {}
-        for _, n in ipairs(nodes) do
-            local vt = parser(n, ft, ts_utils)
-            if vt then
-                local _, col = n:start()
-                local prefix = string.rep(' ', col)
-                table.insert(lines, { { prefix .. vt, opts.highlight } })
-            end
-        end
-        return lines
-    end
-
-    local function text_from_node(node)
-        local vt = parser(node, ft, ts_utils)
-        if vt then
-            return { virt_text = { { vt, opts.highlight } } }
-        end
-        return nil
-    end
-
-    return function(node, nodes)
-        if is_line_ft and vim.tbl_contains(line_targets, node:type()) then
-            if skip_line_ft then
-                return nil
-            end
-
-            local lines = lines_from_nodes(nodes)
-            if #lines > 0 then
-                return { virt_lines = lines }
-            end
-        end
-
-        return text_from_node(node)
-    end
-end
-
-local M = {
-    targets = targets,
-    ignore_root_targets = ignore_root_targets,
-    default_parser = default_parser,
-    default_validator = default_validator,
-    default_resolver = default_resolver,
-}
-
-function M.setup(user_opts)
-    opts = vim.tbl_extend('force', opts, user_opts or {})
-end
-
-function M.show_debug()
-    local ft = parsers.get_buf_lang()
-    local result = find_virtual_text_nodes(function()
-        return true
-    end, ft)
-
-    local values = vim.tbl_values(result)
-    local lines = vim.tbl_keys(result)
-
-    for index, nodes in ipairs(values) do
-        local prefix = string.rep('   ', #values - index)
-        local line = lines[index]
-
-        for _, node in ipairs(nodes) do
-            print(prefix .. line .. ':' .. node:type() .. ' ' .. default_parser(node))
-        end
-    end
-end
-
-function M.show_context()
-    local ft = parsers.get_buf_lang()
-    if vim.tbl_contains(opts.disable_ft, ft) then
-        return
-    end
-
-    local validate = opts.custom_validator or default_validator
-    local resolve = opts.custom_resolver or default_resolver
-    local parse = opts.custom_parser or default_parser
-
-    vim.api.nvim_buf_clear_namespace(0, ns, 0, -1)
-
-    local result = find_virtual_text_nodes(validate, ft)
-    local create_vt = create_virtual_text_factory(parse, ft)
-
-    for line, nodes in pairs(result) do
-        local node = resolve(nodes, ft)
-        local vt = create_vt(node, nodes)
-
-        if vt then
-            vim.api.nvim_buf_set_extmark(0, ns, line, 0, vt)
-        end
-    end
-end
-
-return M
